convert to float and add visualization

+++ include/GraphVisualizer.h
@@ -8,10 +8,10 @@
 template <typename Graph> class ForceDirectedLayout {
 private:
   const Graph &G;
-  std::unordered_map<int, double> NodeIdToX;
-  std::unordered_map<int, double> NodeIdToY;
-  std::unordered_map<int, double> NodeIdToVx;
-  std::unordered_map<int, double> NodeIdToVy;
+  std::unordered_map<int, float> NodeIdToX;
+  std::unordered_map<int, float> NodeIdToY;
+  std::unordered_map<int, float> NodeIdToVx;
+  std::unordered_map<int, float> NodeIdToVy;
 
 public:
   ForceDirectedLayout(const Graph &g) : G(g) {};
@@ -19,7 +19,7 @@ public:
   void initializer() {
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_real_distribution<double> dist(0.0, 800.0);
+    std::uniform_real_distribution<float> dist(0.0f, 800.0f);
     for (int id = 0; id < G.get_node_count(); id++) {
       NodeIdToX[id] = dist(gen);
       NodeIdToY[id] = dist(gen);
@@ -29,12 +29,12 @@ public:
   }
 
   void applyForces() {
-    const double repulsionForce = 1000.0;
-    const double springLength = 100.0;
-    const double springStrength = 0.1;
+    const float repulsionForce = 1000.0;
+    const float springLength = 100.0;
+    const float springStrength = 0.1;
 
-    std::unordered_map<int, double> NodeIdToFx;
-    std::unordered_map<int, double> NodeIdToFy;
+    std::unordered_map<int, float> NodeIdToFx;
+    std::unordered_map<int, float> NodeIdToFy;
     for (int id = 0; id < G.get_node_count(); id++) {
       NodeIdToFx[id] = 0;
       NodeIdToFy[id] = 0;
@@ -42,10 +42,10 @@ public:
 
     for (int i = 0; i < G.get_node_count(); i++) {
       for (int j = i + 1; j < G.get_node_count(); j++) {
-        double dx = NodeIdToX[i] - NodeIdToX[j];
-        double dy = NodeIdToY[i] - NodeIdToY[j];
-        double distance = std::sqrt(dx * dx + dy * dy) + 0.1;
-        double force = repulsionForce / (distance * distance);
+        float dx = NodeIdToX[i] - NodeIdToX[j];
+        float dy = NodeIdToY[i] - NodeIdToY[j];
+        float distance = std::sqrt(dx * dx + dy * dy) + 0.1;
+        float force = repulsionForce / (distance * distance);
 
         NodeIdToFx[i] += (dx / distance) * force;
         NodeIdToFy[i] += (dy / distance) * force;
@@ -56,12 +56,12 @@ public:
 
     for (int u = 0; u < G.get_node_count(); u++) {
       for (int v : G.adj()[u]) {
-        double dx = NodeIdToX[u] - NodeIdToX[v];
-        double dy = NodeIdToY[u] - NodeIdToY[v];
-        double distance = std::sqrt(dx * dx + dy * dy) + 0.1;
+        float dx = NodeIdToX[u] - NodeIdToX[v];
+        float dy = NodeIdToY[u] - NodeIdToY[v];
+        float distance = std::sqrt(dx * dx + dy * dy) + 0.1;
 
-        double displacement = distance - springLength;
-        double force = springStrength * displacement;
+        float displacement = distance - springLength;
+        float force = springStrength * displacement;
 
         NodeIdToFx[u] += (dx / distance) * force;
         NodeIdToFy[u] += (dy / distance) * force;
@@ -77,7 +77,7 @@ public:
   }
 
   void updatePositions() {
-    const double step = 0.01; // Small step size to prevent explosion
+    const float step = 0.01; // Small step size to prevent explosion
     for (int id = 0; id < G.get_node_count(); id++) {
       // Clamp velocity to prevent runaway values
       if (NodeIdToVx[id] > 10) NodeIdToVx[id] = 10;
@@ -98,8 +98,8 @@ public:
       updatePositions();
     }
   }
-  std::unordered_map<int, std::pair<double, double>> getPositions() {
-    std::unordered_map<int, std::pair<double, double>> positions;
+  std::unordered_map<int, std::pair<float, float>> getPositions() {
+    std::unordered_map<int, std::pair<float, float>> positions;
     for (int id = 0; id < G.get_node_count(); id++) {
       positions[id] = {NodeIdToX[id], NodeIdToY[id]};
     }
